+++ src/assets/scss/abstract/_colors.scss
@@ -1,0 +1,10 @@
+$color-text: rgb(234, 235, 236);
+$color-text-feint: #B4B9BC;
+$color-accent: #F7B041;
+$color-border: #656b72;
+$color-border-active: #EAEBEC;
+$color-border-disabled: #3e3e41;
+$color-die-good: rgb(46, 217, 46);
+$color-die-defective: rgb(213, 63, 49);
+$color-die-partial: rgba(249,249,27);
+$color-die-lost: rgba(151,138,129);
