+++ src/assets/scss/base/_forms.scss
@@ -1,0 +1,163 @@
+@use "../abstract/colors" as colors;
+
+.wp-block-create-block-die-yield-calculator {
+	input[type="text"],
+	input[type="number"],
+	select,
+	textarea,
+	input[type="checkbox"],
+	input[type="radio"] {
+		background-color: transparent;
+		border: 1px solid colors.$color-border;
+		color: inherit;
+		transition: border-color 0.2s ease-in-out;
+
+		&:hover {
+			border-color: colors.$color-border-active;
+		}
+
+		&:focus {
+			border-color: colors.$color-accent;
+			outline: none;
+		}
+
+		&:disabled {
+			cursor: not-allowed;
+			border-color: colors.$color-border-disabled;
+			color: colors.$color-text-feint;
+		}
+	}
+
+	input[type="text"],
+	input[type="number"],
+	select,
+	textarea {
+		margin: 0;
+		width: 100%;
+		padding: 15px;
+		border-radius: 10px;
+	}
+
+	select {
+		border-radius: 14px;
+		cursor: pointer;
+		appearance: none;
+		background-image: url("data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iaXNvLTg4NTktMSI/Pgo8IS0tIFVwbG9hZGVkIHRvOiBTVkcgUmVwbywgd3d3LnN2Z3JlcG8uY29tLCBHZW5lcmF0b3I6IFNWRyBSZXBvIE1peGVyIFRvb2xzIC0tPgo8c3ZnIGZpbGw9IiMwMDAwMDAiIGhlaWdodD0iODAwcHgiIHdpZHRoPSI4MDBweCIgdmVyc2lvbj0iMS4xIiBpZD0iTGF5ZXJfMSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiB4bWxuczp4bGluaz0iaHR0cDovL3d3dy53My5vcmcvMTk5OS94bGluayIgCgkgdmlld0JveD0iMCAwIDM4Ni4yNTcgMzg2LjI1NyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+Cjxwb2x5Z29uIHBvaW50cz0iMCw5Ni44NzkgMTkzLjEyOSwyODkuMzc5IDM4Ni4yNTcsOTYuODc5ICIgZmlsbD0iI0VBRUJFQyIvPgo8L3N2Zz4= ");
+		background-position: right 15px center;
+		background-repeat: no-repeat;
+		background-size: 0.75em;
+
+		// To reset Windows from displaying the wrong color
+		option {
+			background-color: initial;
+			color: initial;
+		}
+	}
+
+	input[type="checkbox"],
+	input[type="radio"] {
+		cursor: pointer;
+		margin: 0 0.5em 0 0;
+		appearance: none;
+		-webkit-appearance: none;
+		width: 1.5em;
+		height: 1.5em;
+		outline: none;
+		display: inline-block;
+		position: relative;
+
+		&::before {
+			content: '';
+			opacity: 0;
+			transition: opacity ease-out 0.1s;
+		}
+
+		&:checked::before {
+			opacity: 1;
+		}
+	}
+
+	input[type="checkbox"] {
+		border-radius: 4px;
+
+		&::before {
+			background-image: url("data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyNCIgaGVpZ2h0PSIyNCIgdmlld0JveD0iMCAwIDI0IDI0Ij4KICA8cGF0aCBkPSJNMjAuMjg1IDJsLTExLjI4NSAxMS41NjctNS4yODYtNS4wMTEtMy43MTQgMy43MTYgOSA4LjcyOCAxNS0xNS4yODV6IiBmaWxsPSIjRjdCMDQxIiAvPgo8L3N2Zz4=");
+			background-size: 75%;
+			background-position: center;
+			background-repeat: no-repeat;
+			position: absolute;
+			width: 100%;
+			height: 100%;
+			top: 0;
+			left: 0;
+		}
+
+	}
+
+	input[type="radio"] {
+		border-radius: 100%;
+
+		&::before {
+			width: 50%;
+			height: 50%;
+			background-color: #F7B041;
+			border-radius: 50%;
+			position: absolute;
+			top: 50%;
+			left: 50%;
+			transform: translate(-50%, -50%);
+		}
+	}
+
+	.input-row,
+	.input-row--two-col {
+		margin-bottom: 1rem;
+
+		&--two-col {
+			display: grid;
+			grid-template-columns: 1fr 1fr;
+			column-gap: 1em;
+			align-items: center;
+		}
+	}
+
+	.checkbox {
+		&,
+		&--disabled {
+			margin-bottom: 1em;
+
+			label {
+				display: flex;
+				align-items: center;
+			}
+
+			input[type="checkbox"] {
+				margin-right: 1em;
+			}
+		}
+
+		&--disabled label {
+			color: colors.$color-text-feint;
+			cursor: default;
+		}
+	}
+
+	fieldset {
+		padding: 0;
+		border: 0;
+	}
+
+	.radio-group {
+		margin: 1em 0;
+		display: flex;
+		align-items: center;
+
+	}
+
+	.radio-item {
+		display: flex;
+		align-items: center;
+		line-height: 1;
+		margin: 0 1em 0 0;
+	}
+}
