+++ src/assets/scss/components/_jump-to-results.scss
@@ -1,0 +1,32 @@
+@use "../abstract/colors" as colors;
+
+.wp-block-create-block-die-yield-calculator {
+	.jump-to-results {
+		&,
+		&--hidden {
+			position: fixed;
+			bottom: 2em;
+			left: 50%;
+			padding: 10px 18px;
+			font-size: 16px;
+			cursor: pointer;
+			background-color: colors.$color-accent;
+			color: black;
+			border: none;
+			border-radius: 100px;
+			transform: translateX(-50%);
+			transition: opacity 0.3s ease;
+			border: 0;
+			box-shadow: 0 0.5em 1em rgba(0, 0, 0, 0.75);
+
+			@media (min-width: 768px) {
+				display: none;
+			}
+		}
+
+		&--hidden {
+			opacity: 0;
+			pointer-events: none;
+		}
+	}
+}
