+++ src/assets/scss/components/_result-stats.scss
@@ -1,0 +1,97 @@
+@use "../abstract/colors" as colors;
+
+.wp-block-create-block-die-yield-calculator {
+	.result-stats {
+		display: grid;
+		grid-template-columns: 1fr 1fr;
+	}
+
+	.result-stats__list {
+		list-style: none;
+		padding: 0;
+		margin: 0;
+	}
+
+	.result-stats__result {
+		display: flex;
+		align-items: center;
+		margin-bottom: 0.5em;
+
+		&::before {
+			width: 1rem;
+			margin-right: 0.5rem;
+			line-height: 1rem;
+			text-align: center;
+		}
+
+		&--good-dies,
+		&--defective-dies,
+		&--partial-dies,
+		&--lost-dies,
+		&--total-dies {
+			&::before {
+				content: "■";
+			}
+		}
+
+		&--good-dies::before {
+			color: colors.$color-die-good;
+		}
+
+		&--defective-dies::before {
+			color: colors.$color-die-defective;
+		}
+
+		&--partial-dies::before {
+			color: colors.$color-die-partial;
+		}
+
+		&--lost-dies::before {
+			color: colors.$color-die-lost;
+		}
+
+		&--die-per-reticle::before {
+			content: "▣";
+		}
+
+		&--yield::before {
+			content: "◐";
+			font-size: 1.125em;
+		}
+
+		&--panel-width::before {
+			content: "↔";
+		}
+
+		&--panel-height::before {
+			content: "↕";
+		}
+
+		&--panel-diameter::before {
+			content: "⌀";
+			font-size: 1.5em;
+		}
+
+		&--wafer-area::before {
+			content: "◌";
+		}
+
+		&--die-area::before {
+			content: "⬚";
+		}
+
+		&--shot-count::before {
+			content: "☼";
+		}
+
+		&--reticle-utilization::before {
+			content: "◰";
+			font-size: 1.33333em;
+		}
+
+		&--waste-area::before {
+			content: "⊗";
+			font-size: 1.33333em;
+		}
+	}
+}
