+++ src/assets/scss/components/_wafer-canvas.scss
@@ -1,0 +1,128 @@
+@use "../abstract/colors" as colors;
+@use "../abstract/mixins" as mixins;
+
+.wp-block-create-block-die-yield-calculator {
+	.wafer-canvas {
+		display: inline-block;
+		line-height: 0;
+		margin-bottom: 2em;
+		box-shadow: 0.25em 0.5em 1.5em black;
+		transform-style: preserve-3d;
+		background-color: #b88ecb;
+		touch-action: none;
+		max-width: 100%;
+
+		.wafer-canvas__watermark {
+			position: absolute;
+			top: 0;
+			left: 0;
+			bottom: 0;
+			right: 0;
+			background-image: url("/src/assets/semianalysis-logo-mark-only.png");
+			background-repeat: no-repeat;
+			background-position: 50% 50%;
+			background-size: 60%;
+			opacity: 0.075;
+		}
+
+		.wafer-canvas__mirror-background {
+			@include mixins.rainbow-background;
+
+			border-top: 1px solid rgba(255, 255, 255, 0.5);
+			border-bottom: 1px solid rgba(0, 0, 0, 0.5);
+			position: absolute;
+			top: 0;
+			right: 0;
+			bottom: 0;
+			left: 0;
+		}
+
+		&.wafer-canvas--disc {
+			&,
+			.wafer-canvas__mirror-background {
+				// Use clip path instead of border radius to prevent tilt mouseover being triggered in corners
+				clip-path: ellipse(50% 50% at 50% 50%);
+				transform: rotate(45deg);
+				overflow: hidden;
+			}
+
+			// Display a 1mm v-shaped notch in the bottom of the wafer
+			&::after {
+				content: "";
+				display: block;
+				width: 3px;
+				height: 3px;
+				background-color: #141416;
+				position: absolute;
+				top: 100%;
+				left: 50%;
+				transform: translate(-50%, -50%) rotate(45deg);
+			}
+		}
+
+		.wafer-canvas__die-decorative {
+			max-width: 100%;
+			position: relative;
+		}
+
+		.wafer-canvas__shot-map,
+		.wafer-canvas__edge,
+		.wafer-canvas__die-map {
+			position: absolute;
+			width: 100%;
+			height: 100%;
+			left: 0;
+			top: 0;
+			transition: transform ease-out 0.3s, opacity ease-out 0.3s;
+		}
+
+		&:hover {
+			box-shadow: 0.125em 0.25em 4em rgba(0, 0, 0, 0.75);
+
+			.wafer-canvas__shot-map,
+			.wafer-canvas__die-map,
+			.wafer-canvas__edge {
+				opacity: 0;
+			}
+		}
+
+		.wafer-canvas__message {
+			&,
+			&--error {
+				position: absolute;
+				left: 0;
+				right: 0;
+				top: 50%;
+				text-align: center;
+				transform: translateY(-50%);
+				line-height: 1;
+				color: rgba(0, 0, 0, 0.5);
+				text-shadow: 0 1px #ffffff2e;
+			}
+
+			&--error {
+				color: colors.$color-die-lost;
+			}
+		}
+	}
+
+	.tsmc-logo {
+		position: relative;
+		padding-bottom: 100%;
+		margin-bottom: 2em;
+		border-radius: 100%;
+		overflow: hidden;
+
+		.glare-wrapper {
+			height: 95.6% !important;
+		}
+
+		svg {
+			width: 100%;
+			height: auto;
+			position: absolute;
+			top: 0;
+			left: 0;
+		}
+	}
+}
