+++ src/assets/scss/editor.scss
