+++ src/assets/scss/style.scss
@@ -1,0 +1,8 @@
+@use "base/defaults";
+@use "base/forms";
+@use "layout/index";
+@use "components/jump-to-results";
+@use "components/result-stats";
+@use "components/reticle-canvas";
+@use "components/wafer-canvas";
+
