+++ src/index.js
@@ -12,7 +12,7 @@ import { registerBlockType } from '@wordpress/blocks';
  *
  * @see https://www.npmjs.com/package/@wordpress/scripts#using-css
  */
-import './style.scss';
+import './assets/scss/style.scss';
 
 /**
  * Internal dependencies
